[reorga] Cleaning after pull request
Renaming directories.
DONE mv logger translator
TODO mv server archiver

The logger is the ensemble: translator + archiver

The archiver is for now reading from CSV and dumping in sqlite DB

But it should read from the output of the translator, format the data and send a
formal PUT http request to the http api for database saving.

So that the logger takes care of checking the inputs from the mesh network
and create clean information to archive.