[archiver] Comments correction
The input file is in fact an output file.
And the python script indeed listens to the serial port to save everything into
sqlite.

The thing is that when it starts from and no serial port is present then it
reads data from the sqlite and dumps that into a dataframe.

This is a test phase.